ingestion-service/requirements.txt
fastapi==0.110.3
uvicorn==0.29.0
paho-mqtt==1.6.1
websockets==12.0
